+++ src/inspect_ai/_cli/common.py
@@ -71,13 +71,6 @@ def common_options(func: Callable[..., Any]) -> Callable[..., click.Context]:
         help="Set the display type (defaults to 'full')",
     )
     @click.option(
-        "--no-display-metrics",
-        type=bool,
-        is_flag=True,
-        help="Do not display incremental metrics.",
-        envvar="INSPECT_NO_DISPLAY_METRICS",
-    )
-    @click.option(
         "--no-ansi",
         type=bool,
         is_flag=True,
@@ -114,9 +107,7 @@ def process_common_options(options: CommonOptions) -> None:
         display = "plain"
     else:
         display = options["display"].lower().strip()
-
-    display_metrics = False if options["no_display_metrics"] else True
-    init_display_type(display, display_metrics)
+    init_display_type(display)
 
     # attach debugger if requested
     if options["debug"]: